src/main/java/org/example/portier_digital_admin/auth/UserDetailsServiceImp.java
package org.example.portier_digital_admin.auth;

import org.example.portier_digital_admin.entity.User;
import org.example.portier_digital_admin.service.UserService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
public class UserDetailsServiceImp implements UserDetailsService {
    @Autowired
    private UserService userService;

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        User userByEmail = userService.getByEmail(username);
        return new UserDetailsImp(userByEmail);
    }
}
